refactor: FedimintApi -> crate::FedimintServer, FedimintServer -> consensus::ConsensusServer
